requirements.txt
yfinance>=0.2.48
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b
pyyaml>=6.0
requests>=2.31.0
